Notifications
Summary:
Fixes T11
The author of the post is automatically subscribed to his card (but can always unsubscribe).
The same with everyone posting a comment to a card.

Users can subscribe via the `subscribe/` endpoint under the post detail page (see doc)

On the notification page all unread comments of subscribed posts are displayed.
One sees the area and the id of the post the comment is part of. The client will most likely sort them by post themselves

When the user views the detail page, all comments of that page are marked read. When requesting the comment directly it is not marked read.

Test Plan:
- Post post (and check if subscribed)
- Post comment (and check if subscribed)

- Post comment and retrive notification
- View Post and see notifications disappear

- Mannualy subscribe post
- (Mannualy) unsubscribe post

Reviewers: WyldBot

Reviewed By: WyldBot

Maniphest Tasks: T11

Differential Revision: https://phabricator.wildfyre.net/D15