Use Argon2 as password hasher
Summary:
Until now we used PBKDF2 as the password hasher. Argon2 is stronger,
because it isn't notably faster, when trying to break it with hardware extra designed for breaking that algorythim (e.g. much GPU power),
than on a normal server.

As we don't know the users password, the hash in the database will be updated, when a user logs in the next time.

Test Plan: Unit Tests only.

Reviewers: WyldBot

Reviewed By: WyldBot

Differential Revision: https://phabricator.wildfyre.net/D85