Allow Anonym Posts
Summary:
Posts can be anonym.
An anonym post is still linked to the users account, and the user can still delete it.
It should not be possible (beside using the admin interface), to find out, who posted the anonym post.
The post will however show up in the list of the users own posts.

The admin can filter for anonym posts, and also displays if a post is anonym.

Test Plan:
- Post anonym post
- Post non anonym post
- Delete anonym post
- Check own posts
- View admin list / detail

Reviewers: WyldBot

Reviewed By: WyldBot

Maniphest Tasks: T94

Differential Revision: https://phabricator.wildfyre.net/D112