Email case sensitive workaround
Summary:
Email addresses are safed in the case entered by the user.
When a user requests a list of all his usernames, the users are found
through a case insensitive lookup of the email, and send to the casing used in the request.
When trying to reset a password the mail is always send to the casing assosiated with the account, but searching for the user is still case insensitive

Test Plan:
- Request usrname list
- Reset password

Reviewers: WyldBot

Reviewed By: WyldBot

Maniphest Tasks: T143

Differential Revision: https://phabricator.wildfyre.net/D117