Fix doc Dockerfile
Summary:
The output structure of the documentation generator sphinx changed.
The html output is now in `_build/` and no longer in `_build/html`.

This broke the gerneration of the docker image.

This diff fixes the image, and specifies the version of sphinx directly,
so this does not happen again.

Test Plan: Build the image an run it. Check via browser if the docs work.

Reviewers: WyldBot

Reviewed By: WyldBot

Differential Revision: https://phabricator.wildfyre.net/D233