Disable character validation on User.username
Summary:
The User model had a validation on the username field, that disallowed certain chars.
But because django-rest-framework sadly doesn't use the model validation we now have users with usernames
that are invalid according to the model validation.
These users couldn't be edited in the admin interface (because the admin validates correctly).
This removes the validator form the username field, so that these users can be edited in the admin interface.

Test Plan: Create a user with a space char.

Reviewers: WyldBot

Reviewed By: WyldBot

Differential Revision: https://phabricator.wildfyre.net/D239